# openapi.yaml
openapi: 3.0.3
info:
  title: ApitaCerto API
  version: "1.0"
  description: >
    Backend do dashboard ApitaCerto. Serve dados agregados do Brasileirao
    Serie A (2018 em diante) segmentados por time e arbitro, incluindo o
    Indice de Favorecimento (ver _docs/specs.md secao 6) e a aba Dados
    estatisticos (secao 9).

servers:
  - url: http://localhost:8000

paths:
  /filters:
    get:
      summary: Opcoes disponiveis para os filtros do dashboard
      operationId: getFilters
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
          description: >-
            Quando informado, o campo "teams" volta so com os times que
            jogaram essa temporada (evita listar time rebaixado/de outro
            ano no slicer).
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterOptions"

  /favoritism:
    get:
      summary: >-
        Heatmap time x arbitro pra aba Indice de Favorecimento. Season
        omitido = "Todos", agrega todas as temporadas ja ingeridas (so
        essa aba faz isso -- ganha amostra pro n>=5 do indice).
      operationId: getFavoritism
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [heatmap]
                properties:
                  heatmap:
                    type: array
                    items:
                      $ref: "#/components/schemas/HeatmapCell"

  /season-overview:
    get:
      summary: >-
        KPIs de temporada (rodada atual, cartoes/jogo, aproveitamento do
        mandante, gols/jogo) e rankings de arbitro (mais cartao, vies de
        mandante) pra tela Dashboard.
      operationId: getSeasonOverview
      parameters:
        - name: season
          in: query
          required: true
          schema:
            type: integer
            minimum: 2018
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeasonOverview"

  /statistics:
    get:
      summary: >-
        Aba Dados estatisticos (spec secao 9): relatorio pre-calculado offline
        por scripts/compute_stats.py e so lido aqui. Season omitido = todas
        as temporadas juntas.
      operationId: getStatistics
      parameters:
        - name: season
          in: query
          required: false
          schema:
            type: integer
            minimum: 2018
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatisticsReport"
        "404":
          description: Relatorio ainda nao calculado para esse recorte

  /dashboard:
    get:
      summary: Dados agregados do dashboard para o recorte de filtros dado
      operationId: getDashboard
      parameters:
        - name: season
          in: query
          required: true
          schema:
            type: integer
            minimum: 2018
        - name: team
          in: query
          required: false
          schema:
            type: string
        - name: referee
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"

components:
  schemas:
    FilterOptions:
      type: object
      required: [teams, referees, seasons]
      properties:
        teams:
          type: array
          items: { type: string }
        referees:
          type: array
          items: { type: string }
        seasons:
          type: array
          items: { type: integer }

    Kpis:
      type: object
      required:
        [games, wins, draws, losses, winRatePct, goalsFor, goalsAgainst, yellow, red]
      properties:
        games: { type: integer }
        wins: { type: integer }
        draws: { type: integer }
        losses: { type: integer }
        winRatePct: { type: number }
        goalsFor: { type: integer }
        goalsAgainst: { type: integer }
        yellow: { type: integer }
        red: { type: integer }

    TimeseriesPoint:
      type: object
      required: [round, winRatePct, yellow, red]
      properties:
        round: { type: integer }
        winRatePct: { type: number }
        yellow: { type: integer }
        red: { type: integer }

    HeatmapCell:
      type: object
      required: [team, referee, n, insufficientSample]
      properties:
        team: { type: string }
        referee: { type: string }
        n: { type: integer, description: "jogos do par time-arbitro na temporada" }
        wins: { type: integer }
        draws: { type: integer }
        losses: { type: integer }
        yellow: { type: integer }
        red: { type: integer }
        insufficientSample:
          type: boolean
          description: "true quando n < 5 (piso de amostra, spec secao 6)"
        index:
          type: number
          nullable: true
          description: "Indice de Favorecimento; null quando insufficientSample"

    RefereeRanking:
      type: object
      required: [name, games]
      properties:
        name: { type: string }
        games: { type: integer }
        cardsPerGame: { type: number, description: "so em mostCardsReferees" }
        barPct: { type: number, description: "cardsPerGame normalizado no maximo da lista, 0-100; so em mostCardsReferees" }
        bias: { type: number, description: "cartoes/jogo mandante menos visitante; negativo = visitante leva mais; so em homeBiasReferees" }

    SeasonOverview:
      type: object
      required: [currentRound, gamesPlayed, cardsPerGame, homeWinPct, goalsPerGame, mostCardsReferees, homeBiasReferees, allReferees]
      properties:
        currentRound: { type: integer }
        gamesPlayed: { type: integer }
        cardsPerGame: { type: number }
        homeWinPct: { type: number }
        goalsPerGame: { type: number }
        mostCardsReferees:
          type: array
          items: { $ref: "#/components/schemas/RefereeRanking" }
          description: "top 5 por cardsPerGame, pra tela Dashboard"
        homeBiasReferees:
          type: array
          items: { $ref: "#/components/schemas/RefereeRanking" }
          description: "top 4 por vies (mais negativo primeiro), pra tela Dashboard"
        allReferees:
          type: array
          items: { $ref: "#/components/schemas/RefereeRanking" }
          description: "todo arbitro elegivel (>= piso de amostra) com cardsPerGame e bias, pra tela Arbitros"

    DataCompleteness:
      type: object
      required: [isReal, fixtures, fixturesWithCards, lastUpdated]
      properties:
        isReal:
          type: boolean
          description: "false = temporada ainda sem nenhum jogo real ingerido, dashboard mostrando dado sintetico (mock)"
        fixtures:
          type: integer
          description: "jogos reais com placar no banco pra essa temporada"
        fixturesWithCards:
          type: integer
          description: "desses, quantos ja tiveram cartao/gol processado"
        lastUpdated:
          type: string
          format: date-time
          nullable: true
          description: "ISO 8601 UTC da ultima rodada de scraping processada pra essa temporada; null se isReal=false"

    EvidenceLevel:
      type: string
      enum: [forte, fraco, acaso, "sem comparação"]
      description: >-
        forte = q <= 0,10 (Benjamini-Hochberg); fraco = p < 0,01 sem
        sobreviver a correcao; acaso = resto (spec 9.2)

    PairTest:
      type: object
      required: [team, referee, n, z, p, q, level]
      properties:
        team: { type: string }
        referee: { type: string }
        n: { type: integer, description: "jogos do par no recorte" }
        observed: { type: number }
        expected: { type: number }
        ratio: { type: number, nullable: true, description: "observado / esperado (so contagens)" }
        perGame: { type: number, description: "(observado - esperado) por jogo" }
        easeWithReferee: { type: number, description: "so em ease: xPts medio dos jogos com o arbitro" }
        easeTeam: { type: number, description: "so em ease: xPts medio esperado sob sorteio" }
        sameCategory: { type: object, description: "so em ease: mesmo teste contra arbitros da mesma categoria" }
        z: { type: number }
        p: { type: number }
        q: { type: number }
        level: { $ref: "#/components/schemas/EvidenceLevel" }

    PairAnalysis:
      type: object
      required: [floor, points, funnel, summary]
      properties:
        floor: { type: integer, description: "piso de jogos pra entrar no teste" }
        points:
          type: array
          items: { $ref: "#/components/schemas/PairTest" }
        funnel:
          type: array
          description: "linhas do grafico de funil (x, lo95, hi95, lo998, hi998)"
          items: { type: object }
        summary: { type: object, description: "testados, fortes, fracos, esperado por acaso" }

    StatisticsReport:
      type: object
      required: [key, seasons, methodVersion, overview, pairs, escala, league, crossSeason, hypotheses, computedAt]
      properties:
        key: { type: string, description: "temporada ou 'all'" }
        season: { type: integer, nullable: true }
        seasons: { type: array, items: { type: integer } }
        partialSeasons: { type: array, items: { type: integer }, description: "temporadas em andamento no recorte" }
        methodVersion: { type: integer }
        computedAt: { type: string, format: date-time }
        dataVersion: { type: string }
        overview: { type: object }
        refereeStrictness: { type: object, description: "L2 -- rigor ajustado com encolhimento e vies de mandante vs liga" }
        pairs:
          type: object
          properties:
            cards: { $ref: "#/components/schemas/PairAnalysis" }
            rivalCards: { $ref: "#/components/schemas/PairAnalysis" }
            points: { $ref: "#/components/schemas/PairAnalysis" }
            ease: { $ref: "#/components/schemas/PairAnalysis" }
        escala: { type: object, description: "E1 federacao, E2 categoria x importancia, E3 concentracao" }
        league: { type: object, description: "L1 -- mandante vs visitante por temporada" }
        crossSeason: { type: object, description: "A5 -- mesmo par em temporadas diferentes" }
        hypotheses: { type: array, items: { type: object }, description: "H1-H4 pre-registradas" }

    DashboardResponse:
      type: object
      required: [kpis, timeseries, heatmap, dataCompleteness]
      properties:
        kpis:
          $ref: "#/components/schemas/Kpis"
        timeseries:
          type: array
          items:
            $ref: "#/components/schemas/TimeseriesPoint"
        heatmap:
          type: array
          items:
            $ref: "#/components/schemas/HeatmapCell"
        dataCompleteness:
          $ref: "#/components/schemas/DataCompleteness"